[CARD-12] refactor code of store
