+++ src/components/formSignUp/formSignUp.tsx
@@ -3,8 +3,7 @@ import { useHistory } from "react-router-dom";
 import { PropsForm } from "@/types/types";
 import { usersRegistration } from "@/services/dataService";
 import { useDispatch, useSelector } from "react-redux";
-import { authLogInAction, authLogOutAction } from "@/store/actionCreators/authActions";
-import { userNameAction } from "@/store/actionCreators/userNameAction";
+import { authLogInAction, authLogOutAction, userNameAction } from "@/store/actionCreators/authActions";
 import mockServerHelper from "webpack-mock-server/lib/mockServerHelper";
 import { validateLogin, validatePassword, validateRepeatPassword } from "@/utils/validation";
 import { removeAllItems } from "@/store/actionCreators/cartActions";
