--- src/store/actionCreators/userNameAction.tsx
@@ -1,3 +1,0 @@
-import { USER_NAME } from "@/store/actions";
-
-export const userNameAction = (payload: string) => ({ type: USER_NAME, payload });
