+++ src/store/actions/adminAction.tsx
@@ -1,0 +1,4 @@
+export const CREATE_NEW_GAME = "CREATE_NEW_GAME";
+export const EDIT_GAME = "EDIT_GAME";
+export const DELETE_SET_SUCCESS = "DELETE_SET_SUCCESS";
+export const LOAD_GAME = "LOAD_GAME";
