+++ src/store/actions/authAction.tsx
@@ -1,0 +1,5 @@
+export const LOG_IN = "LOG_IN";
+export const LOG_OUT = "LOG_OUT";
+export const USER_NAME = "USER_NAME";
+export const OPEN_MODAL = "OPEN_MODAL";
+export const CLOSE_MODAL = "CLOSE_MODAL";
