+++ src/store/actions/cartAction.tsx
@@ -1,14 +1,3 @@
-export const LOG_IN = "LOG_IN";
-export const LOG_OUT = "LOG_OUT";
-export const USER_NAME = "USER_NAME";
-export const OPEN_MODAL = "OPEN_MODAL";
-export const CLOSE_MODAL = "CLOSE_MODAL";
-
-export const CREATE_NEW_GAME = "CREATE_NEW_GAME";
-export const EDIT_GAME = "EDIT_GAME";
-export const DELETE_SET_SUCCESS = "DELETE_SET_SUCCESS";
-export const LOAD_GAME = "LOAD_GAME";
-
 export const ADD_TO_CART = "ADD_TO_CART";
 export const ADJUST_ITEM_QTY = "ADJUST_ITEM_QTY";
 export const REMOVE_FROM_CART = "REMOVE_FROM_CART";
