+++ src/store/reducers/authReducer.tsx
@@ -1,13 +1,21 @@
 import initialState from "@/store/initialState";
-import { ActionTypes, IAction, IInitialState } from "@/types/types";
+import { IInitialState } from "@/types/types";
+import { ActionTypes, IAction } from "../types/authTypes";
 
 const authReducer = (state = initialState, action: IAction): IInitialState => {
+  console.log(action.type);
   switch (action.type) {
     case ActionTypes.LOG_IN:
       return { ...state, auth: true };
     case ActionTypes.LOG_OUT:
       localStorage.clear();
       return { ...state, auth: false };
+    case ActionTypes.OPEN_MODAL:
+      return { ...state, modal: true };
+    case ActionTypes.CLOSE_MODAL:
+      return { ...state, modal: false };
+    case ActionTypes.USER_NAME:
+      return { ...state, userName: action.payload };
     default:
       return state;
   }
