+++ src/types/types.tsx
@@ -95,165 +95,6 @@ export interface IInitialState {
   total: number;
 }
 
-export interface saveProfile {
-  login: string;
-  password: string;
-  description: string;
-  photo: string;
-  id: number;
-}
-
-export enum ActionTypes {
-  LOG_IN = "LOG_IN",
-  LOG_OUT = "LOG_OUT",
-  USER_NAME = "USER_NAME",
-  OPEN_MODAL = "OPEN_MODAL",
-  CLOSE_MODAL = "CLOSE_MODAL",
-}
-
-export enum ActionAdminTypes {
-  CREATE_NEW_GAME = "CREATE_NEW_GAME",
-  EDIT_GAME = "EDIT_GAME",
-  DELETE_SET_SUCCESS = "DELETE_SET_SUCCESS",
-  LOAD_GAME = "LOAD_GAME",
-}
-
-export enum ActionCartTypes {
-  ADD_TO_CART = "ADD_TO_CART",
-  ADJUST_ITEM_QTY = "ADJUST_ITEM_QTY",
-  REMOVE_FROM_CART = "REMOVE_FROM_CART",
-  REMOVE_ALL_ITEM = "REMOVE_ALL_ITEM",
-  TOTAL_ITEMS = "TOTAL_ITEMS",
-  LOCAL_STORAGE_CART = "LOCAL_STORAGE_CART",
-  LOAD_CART_PRODUCTS = "LOAD_CART_PRODUCTS",
-  UPDATE_CART_PRODUCTS = "UPDATE_CART_PRODUCTS",
-}
-
-interface loadCartProductsAction {
-  type: ActionCartTypes.LOAD_CART_PRODUCTS;
-  payload: {
-    productsCartList: dataItems[];
-    id?: number;
-  };
-}
-
-interface updateCartProductsAction {
-  type: ActionCartTypes.UPDATE_CART_PRODUCTS;
-  payload: {
-    newGameID: number;
-    newGame: dataItems;
-    id?: number;
-  };
-}
-
-
-interface loadGames {
-  type: ActionAdminTypes.LOAD_GAME;
-  payload: {
-    products: dataItems[];
-  };
-}
-
-interface deleteSetSuccess {
-  type: ActionAdminTypes.DELETE_SET_SUCCESS;
-  payload: {
-    id: number;
-  };
-}
-
-interface editGames {
-  type: ActionAdminTypes.EDIT_GAME;
-  payload: {
-    game: dataItems;
-  };
-}
-
-interface createGame {
-  type: ActionAdminTypes.CREATE_NEW_GAME;
-  payload: {
-    game: dataItems;
-  };
-}
-
-export type IActionAdmin = createGame | editGames | deleteSetSuccess | loadGames;
-
-interface addCartFromLS {
-  type: ActionCartTypes.LOCAL_STORAGE_CART;
-  payload: {
-    savedGames: ICart[];
-    id?: number;
-  };
-}
-
-interface totalItemsCart {
-  type: ActionCartTypes.TOTAL_ITEMS;
-  payload: {
-    num: number;
-    id?: number;
-  };
-}
-
-interface addToCart {
-  type: ActionCartTypes.ADD_TO_CART;
-  payload: {
-    id: number;
-  };
-}
-
-interface removeFromCart {
-  type: ActionCartTypes.REMOVE_FROM_CART;
-  payload: {
-    id: number;
-  };
-}
-
-interface adjustItemQty {
-  type: ActionCartTypes.ADJUST_ITEM_QTY;
-  payload: {
-    id: number;
-    qty: number;
-  };
-}
-
-interface removeAllItems {
-  type: ActionCartTypes.REMOVE_ALL_ITEM;
-}
-
-export type IActionCart =
-  | addToCart
-  | adjustItemQty
-  | removeFromCart
-  | removeAllItems
-  | totalItemsCart
-  | addCartFromLS
-  | loadCartProductsAction
-  | updateCartProductsAction;
-
-interface LoginTypeAction {
-  type: ActionTypes.LOG_IN;
-}
-interface logoutTypeAction {
-  type: ActionTypes.LOG_OUT;
-}
-interface openModalTypeAction {
-  type: ActionTypes.OPEN_MODAL;
-}
-interface closeModalTypeAction {
-  type: ActionTypes.CLOSE_MODAL;
-}
-
-interface userNameTypeAction {
-  type: ActionTypes.USER_NAME;
-  payload: string;
-}
-
-export type IAction =
-  | LoginTypeAction
-  | logoutTypeAction
-  | openModalTypeAction
-  | closeModalTypeAction
-  | userNameTypeAction;
-
 export interface IFilterState {
   genre: string;
   age: string;
